requirements.txt
flask_restplus
Flask
marshmallow
arrow
flask-cors
flask-marshmallow
webargs
apispec>=1.0.0b1
apispec_webframeworks
flask_injector
requests
timeout-decorator
psycopg2-binary
pytest-flask
git+https://github.com/jbkoh/brick_data.git@master
pyaml
Django==1.10.6
django-braces==1.11.0
django-filters==0.2.1
django-guardian==1.4.6
django-jinja==2.2.2
django-oauth-toolkit==0.12.0
django-oauth2-provider==0.2.6.1
django-registration==2.2
djangorestframework==3.6.2
python-building-depot==0.2
